+++ tests/index.js
@@ -14,10 +14,12 @@ import * as worker from './worker.tests.js'
 import * as permissions from './permissions.tests.js'
 import * as permissionsEnforcement from './permissionsEnforcement.tests.js'
 import * as logger from './logger.tests.js'
+import * as config from './config.tests.js'
 import { runTests } from 'lib0/testing'
 
 runTests({
   logger,
+  config,
   permissions,
   permissionsEnforcement,
   computeWorker,